% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateClimate.R
\name{aggregateClimate}
\alias{aggregateClimate}
\title{Aggregate climate parameters in time and by station}
\usage{
aggregateClimate(dat, dts, plotID, variable, fun, agg = "month")
}
\arguments{
\item{dat}{data.frame including Dates, climate parameters for different plots}

\item{dts}{Character. Indicates which column of dat includes the Dates}

\item{plotID}{Character. Indicates which column of dat includes the Plot identifier}

\item{variable}{Character. Indicates which column of dat includes the variable to be aggregated}

\item{fun}{Character. How to aggregate the data. e.g. "mean" or "sum".}

\item{agg}{Character. defines the aggregation level. Either "week","month" or "year"}
}
\value{
A data.frame with aggregated climate information
}
\description{
Aggregate climate parameters in time and by station
}
\examples{
#Calculate GDD from MODIS day/night data for several plots and aggregate it per month
dat <- get(load(system.file("extdata","ClimateData.RData",package="metTools")))
dat$GDD <- GDD(tmax=dat$tday,tmin=dat$tnight)
GDD_agg <- aggregateClimate(dat,"Date","Plot","GDD","sum")
}
\seealso{
\code{\link{GDD}}
}
\author{
Hanna Meyer
}
