% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amedas.R
\name{amedas.stations}
\alias{amedas.stations}
\title{generate AMeDAS station list in a data.frame}
\usage{
amedas.stations()
}
\value{
df
}
\description{
generate AMeDAS station list in a data.frame
}
\examples{
df <- amedas.stations()
}
