% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{calc.condtemp}
\alias{calc.condtemp}
\title{dew point temperature}
\usage{
calc.condtemp(T, e)
}
\arguments{
\item{T}{temperature}

\item{e}{vapour pressure Pa}
}
\value{
condensation temperature K
}
\description{
Calculate dew point temperature
Bolton (1980)
}
\examples{
t.cond <- calc.condtemp(300, 2000)
}
