% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{calc.es}
\alias{calc.es}
\title{saturation vapour pressure}
\usage{
calc.es(T)
}
\arguments{
\item{T}{temperature}
}
\value{
saturation vapour pressure Pa
}
\description{
Calculate saturation vapour pressure
based on the formula used by WMO and JMA
}
\examples{
es <- calc.es(300)
}
