% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{calc.thetae}
\alias{calc.thetae}
\title{equivalent potential temperature}
\usage{
calc.thetae(T, e, p)
}
\arguments{
\item{T}{temperature K}

\item{e}{vapour pressure Pa}

\item{p}{pressure Pa}
}
\value{
equivalent potential temperature K
}
\description{
Calculate equivalent potential temperature
Bolton (1980)
}
\examples{
thetae <- calc.thetae(300, 1000, 850e2)
}
