% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{calc.thetaes}
\alias{calc.thetaes}
\title{Calculate saturated equivalent potential temperature
Bolton (1980)}
\usage{
calc.thetaes(T, p)
}
\arguments{
\item{T}{temperature K}

\item{p}{pressure Pa}
}
\value{
saturated equivalent potential temperature K
}
\description{
Calculate saturated equivalent potential temperature
Bolton (1980)
}
\examples{
thetaes <- calc.thetaes(300, 1000e2)
}
