% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMetAnnual.R
\name{compMetAnnual}
\alias{compMetAnnual}
\title{Compute annual meteorological characteristics}
\usage{
compMetAnnual(data, gid, scol, mcol, ...)
}
\arguments{
\item{data}{Data frame containing monthly meteorological variables}

\item{gid}{Grouping ID, e.g. plot and year for aggregation}

\item{scol}{Columns for}
}
\value{
A \code{data.frame} object.
}
\description{
Compute annual meteorological characteristics.
}
\examples{
\\notrun{
}

}
\references{
NONE
}
\author{
Thomas Nauss
}
