% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{cshift}
\alias{cshift}
\title{Cyclic Index Shifter}
\usage{
cshift(n, k)
}
\arguments{
\item{n}{An integer scalar representing the maximum size of the grid dimension.}

\item{k}{An integer scalar representing the size and direction of the index shift.}
}
\value{
A vector of length \code{n} containing the shifted indices wrapped within range \code{1:n}.
}
\description{
A utility function to compute shifted matrix indices for periodic/cyclic
boundary conditions.
}
\keyword{internal}
