% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jma_besttrack.R
\name{decode_data}
\alias{decode_data}
\title{Decode RSMC Tokyo Best Track Data Lines}
\usage{
decode_data(data)
}
\arguments{
\item{data}{A character vector where each element is a data line for a specific tropical cyclone.}
}
\value{
A data frame containing parsed track data, including time, location, intensity, and wind radii.
}
\description{
Parses a character vector of best track data lines into a data frame containing storm track details.
}
\keyword{internal}
