% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amedas.R
\name{dir2deg}
\alias{dir2deg}
\title{convert wind directions to degrees}
\usage{
dir2deg(x)
}
\arguments{
\item{x}{wind directions}
}
\value{
degreee
}
\description{
convert wind directions to degrees
}
\examples{
df <- dir2deg(c("北"))
}
