% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{e2q}
\alias{e2q}
\title{vapour pressure to specific humidity}
\usage{
e2q(e, p)
}
\arguments{
\item{e}{vapour pressure Pa| hPa}

\item{p}{air pressure Pa | hPa (must be consistent with e)}
}
\value{
specific humidity kg/kg
}
\description{
Convert vapour pressure to specific humidity
}
\examples{
q <- e2q(20, 1000)
q <- e2q(20e2, 1000e2)
}
