% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{e2w}
\alias{e2w}
\title{vapour pressure to mixing ratio}
\usage{
e2w(e, p)
}
\arguments{
\item{e}{vapour pressure Pa| hPa}

\item{p}{air pressure Pa | hPa (must be consistent with e)}
}
\value{
mixing ratio kg/kg
}
\description{
Convert vapour pressure to mixing ratio
}
\examples{
q <- e2w(20, 1000)
q <- e2w(20e2, 1000e2)
}
