% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compEcoTemp.R
\name{GDD}
\alias{GDD}
\alias{Gsum}
\alias{Coldsum}
\title{Calculate ecologically relevant temperature parameters}
\usage{
GDD(tmax, tmin, base = 10)

Gsum(tmean, base = 5)

Coldsum(tmean, base = 0)
}
\arguments{
\item{tmax}{Numeric. Daily max temperatures}

\item{tmin}{Numeric. Daily min temperatures}

\item{base}{Numeric. Base temperature to which the parameters will be calculated}

\item{tmean}{Numeric. Daily mean temperatures}
}
\value{
A Vector with calculated parametes with the same length as the input
}
\description{
Calculate ecologically relevant temperature parameters
}
\note{
See \code{\link{aggregateClimate}} to aggregate the results to weeks, months, years
}
\examples{
GDD(tmax=c(5,6,10,17),tmin=c(4,5,6,13))
Gsum(tmean=c(5,11,10,17))
}
\seealso{
\code{\link{aggregateClimate}}
}
\author{
Hanna Meyer
}
