% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igra2.R
\name{igra2_derived}
\alias{igra2_derived}
\title{download IGRA2 derived station data}
\usage{
igra2_derived(id, method = "libcurl")
}
\arguments{
\item{id}{station id}

\item{method}{download method}
}
\description{
download IGRA2 station derived data for the full period
}
