% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igra2.R
\name{igra2_full}
\alias{igra2_full}
\title{download IGRA2 station data}
\usage{
igra2_full(id, method = "libcurl")
}
\arguments{
\item{id}{station id}

\item{method}{download method}
}
\description{
download IGRA2 station data for the full period
}
