% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igra2.R
\name{igra2_monthly}
\alias{igra2_monthly}
\title{download IGRA2 monthly mean}
\usage{
igra2_monthly(var, hh, method = "libcurl")
}
\arguments{
\item{var}{one of "ghgt", "temp", "uwnd", "vapr", "vwnd"}

\item{hh}{00 or 12}

\item{method}{download method}
}
\description{
download IGRA2 monthly mean for specified hour and variable
}
