% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igra2.R
\name{igra2_y2d}
\alias{igra2_y2d}
\title{download IGRA2 yearly station data}
\usage{
igra2_y2d(id, year = NULL, method = "libcurl")
}
\arguments{
\item{id}{station id}

\item{year}{year}

\item{method}{download method}
}
\description{
download IGRA2 station data for the last year
}
