% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metTools-package.R
\docType{package}
\name{metTools-package}
\alias{metTools-package}
\alias{metToolsToolspackage}
\title{Tools for meteorological analysis}
\description{
Tools for meteorological analysis
}
\details{
We provide a set of functions usefull for some of our common meteorological
analysis studies.
}
\author{
Thomas Nauss
}
\keyword{package}
