% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{q2e}
\alias{q2e}
\title{specific humidity to vapour pressure}
\usage{
q2e(q, p)
}
\arguments{
\item{q}{specific humidity kg/kg}

\item{p}{air pressure Pa | hPa}
}
\value{
vapour pressure in the same unit as p
}
\description{
Convert specific humidity vapour pressure
}
\examples{
e.hPa <- q2e(0.01, 1000)
e.Pa <- q2e(0.01, 1000e2)
}
