% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{q2ttd}
\alias{q2ttd}
\title{specific humidity to dew point depression}
\usage{
q2ttd(q, T, p)
}
\arguments{
\item{q}{specific humidity kg/kg}

\item{T}{temperature}

\item{p}{pressure Pa}
}
\value{
dew point depression K
}
\description{
Calculate dew point depression from specific humidity
}
\examples{
ttd <- q2ttd(0.01, 300, 1000e2)
}
