% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMetData.R
\name{readMetData}
\alias{readMetData}
\title{Read meteorological data}
\usage{
readMetData(filepath, ...)
}
\arguments{
\item{filepath}{Full path and name of the csv file containing the data.}
}
\value{
A \code{data.frame} object.
}
\description{
Read meteorological data from files provided in the format of our
climate database.
}
\examples{
\\notrun{
}

}
\references{
NONE
}
\author{
Thomas Nauss
}
