% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{ttd2q}
\alias{ttd2q}
\title{dew point depression to specific humidity}
\usage{
ttd2q(ttd, T, p)
}
\arguments{
\item{ttd}{dew point depression (T - Td)}

\item{T}{temperature}

\item{p}{pressure Pa}
}
\value{
specific humidity kg/kg
}
\description{
Calculate specific humidity from dew point depression
}
\examples{
q <- ttd2q(12, 300, 1000e2)
}
