% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jma_besttrack.R
\name{decode_bst}
\alias{decode_bst}
\title{Decode JMA Best Track Connection}
\usage{
decode_bst(con)
}
\arguments{
\item{con}{A connection to the extracted text file.}
}
\value{
A named list of data frames, where each element contains track data for a tropical cyclone, indexed by its international ID.
}
\description{
Reads and parses the text connection extracted from the Best Track zip file.
}
\keyword{internal}
