% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{div}
\alias{div}
\title{Spatial Divergence of a 2D Vector Field}
\usage{
div(u, v, hx, hy = hx, cyclic = FALSE)
}
\arguments{
\item{u}{A numeric matrix representing the horizontal/longitudinal components of the vector field.}

\item{v}{A numeric matrix representing the vertical/latitudinal components of the vector field.}

\item{hx}{A numeric scalar representing the grid spacing in the x-direction.}

\item{hy}{A numeric scalar representing the grid spacing in the y-direction. 
Defaults to \code{hx}.}

\item{cyclic}{A logical scalar. Set to \code{TRUE} if the field domain wraps periodically. 
Defaults to \code{FALSE}.}
}
\value{
A numeric matrix matching the input vector field grid dimensions containing 
  the computed spatial divergence.
}
\description{
Calculates the scalar divergence field \code{div(u,v)} 
\eqn{\nabla \cdot \mathbf{v} = \frac{\partial u}{\partial x} + \frac{\partial v}{\partial y}} 
from its coordinate vector components.
}
