% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physcon}
\alias{physcon}
\title{physical constants of air and water}
\format{
## `physicon`
physical constants of air and water
\describe{
  \item{air.rd}{gas constant of dry air, J/deg/kg}
  \item{air.cp}{specific heat at constant pressure, J/deg/kg}
  \item{air.cv}{specific heat at constant volume, J/deg/kg}
  \item{air.kappa}{ratio of gas constant to specific heat at constant pressure of dry air}
  \item{air.gamma}{ratio of specific heat at constant pressure to that at constant volume of dry air}
  \item{air.pr}{reference pressure, Pa}
  \item{water.t0}{freezing temperature, K}
  \item{water.rv}{gas constant of water vapour, J/deg/kg}
  \item{water.lv0}{latent heat of vapouriztion at 0C , J/kg}
  \item{water.eps}{ration of gas constant of dry air to that of water vapour}
  \item{earth.grav}{acceleration due to gravity, m/s^2}
  \item{earth.dayinsec}{length of a day, seconds}
  \item{earth.omega}{angular velocity of the Earth, rad/s}
}
}
\source{
{Wallace and Hobbs (2006)}
}
\usage{
physcon
}
\description{
physical constants of air and water
}
\keyword{datasets}
