% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist.R
\name{rh2q}
\alias{rh2q}
\title{relative humidity to specific humidity}
\usage{
rh2q(rh, T, p)
}
\arguments{
\item{rh}{relative humidity percent}

\item{T}{temperature K}

\item{p}{pressure Pa}
}
\value{
specific humidity kg/kg
}
\description{
Calculate specific humidity from relative humidity
}
\examples{
q <- rh2q(50, 300, 1000e2)
}
